requirements.txt
python-telegram-bot
pytesseract
openai
selenium
pydub
pandas